Make cloud id and url both optional and add tests
